Move references to index
