+++ src/components/NavBar.tsx
@@ -1,7 +1,6 @@
-import Link from "next/link"
 import { useRouter } from "next/router"
-import { useState } from "react"
-type NavBarProps = {
+
+export type NavBarProps = {
   navOption: string,
   updateNavOption: (option: string) => void, 
 }
